+++ example/cached_network.dart
@@ -2,5 +2,5 @@ import 'package:cached_network/cached_network.dart';
 
 void main() async {
   var network = CachedNetwork();
-  await network.fetch('https://google.com');
+  await network.request('https://google.com');
 }