+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+
+- rename `fetch` into `request` _(This is an **incompatible** update)_
+- reacquire from network when cache has no content
+
 ## 0.0.3
 
 - add dio param when construct instance